Use queued job for "unsearching" when Scout queue is enabled (#471)
* Queue scout removals

* Fix code doc

* Undo new line

* Fix style

* Update src/Jobs/RemoveFromSearch.php


* Update src/Jobs/RemoveFromSearch.php

+++ src/Searchable.php
@@ -5,6 +5,7 @@ namespace Laravel\Scout;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Collection as BaseCollection;
 use Laravel\Scout\Jobs\MakeSearchable;
+use Laravel\Scout\Jobs\RemoveFromSearch;
 
 trait Searchable
 {
@@ -80,7 +81,13 @@ trait Searchable
             return;
         }
 
-        return $models->first()->searchableUsing()->delete($models);
+        if (! config('scout.queue')) {
+            return $models->first()->searchableUsing()->delete($models);
+        }
+
+        dispatch(new RemoveFromSearch($models))
+            ->onQueue($models->first()->syncWithSearchUsingQueue())
+            ->onConnection($models->first()->syncWithSearchUsing());
     }
 
     /**
